Move UnitGroup to OpenTabletDriver.UX.Controls.Generic
