+++ OpenTabletDriver.UX/Controls/Output/Area/RotationAreaEditor.cs
@@ -1,4 +1,5 @@
 using Eto.Forms;
+using OpenTabletDriver.UX.Controls.Generic;
 using OpenTabletDriver.UX.Controls.Generic.Text;
 using OpenTabletDriver.UX.Controls.Utilities;
